main/pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "pymongo>=4.6.1",
    "fastapi>=0.108.0",
    "uvicorn>=0.25.0",
    "authlib>=1.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "itsdangerous>=2.1.2",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"
[tool.pdm.scripts]
dev = "uvicorn src.main:app --reload"